udpate feature table package name
