+++ data_processing/radiology/unpack_images/config.yaml.template
