+++ tests/data_processing/radiology/unpack_images/data.yaml
@@ -9,7 +9,7 @@ DATA_TYPE: FEATURE
 DATASET_NAME: unpack
 
 # local path to unpack pngs.
-DESTINATION_PATH: tests/data_processing/testdata/data/unpacked_pngs
+DESTINATION_PATH: tests/data_processing/radiology/testdata/unpacked_pngs
 
 # field to unpack dicom or overlay
 COLUMN_NAME: dicom
